+++ src/devices/philips.ts
@@ -13,17 +13,17 @@ const ea = exposes.access;
 
 export const definitions: DefinitionWithExtend[] = [
     {
-        zigbeeModel: ["LCX029"],
+        zigbeeModel: ["LCX029", "LCX028"],
         model: "929004581801",
         vendor: "Philips",
-        description: "Hue Festivia globe outdoor string lights",
+        description: "Hue Festavia globe outdoor string lights",
         extend: [philips.m.light({colorTemp: {range: [50, 1000]}, color: {modes: ["xy", "hs"], enhancedHue: true}})],
     },
     {
         zigbeeModel: ["LCX030"],
         model: "929004284702",
         vendor: "Philips",
-        description: "Hue Festavia permanent string lights",
+        description: "Hue Festavia permanent outdoor string lights",
         extend: [philips.m.light({colorTemp: {range: [50, 1000]}, color: {modes: ["xy", "hs"], enhancedHue: true}})],
     },
     {